+++ src/components/DatePicker/DatePicker.story.js
@@ -1,3 +1,4 @@
+/* eslint-disable react/jsx-no-bind */
 import React from 'react';
 import styled from 'styled-components';
 import moment from 'moment';
@@ -64,13 +65,68 @@ stories.add('Default single date picker', (() => (
       />
     </Container>
   )))
-  .add('With past date selectable', (() => (
+  .add('Allow any dates', (() => (
     <Container>
       <DatePicker
         inputId="date-picker-id"
         initialDate={ null }
         isFocused={ boolean('isFocused', true) }
+        isFutureDateSelectable={ boolean('isFutureDateSelectable', true) }
         isPastDateSelectable={ boolean('isPastDateSelectable', true) }
+        onDateChange={ action('onDateChange') }
+      />
+    </Container>
+  )))
+  .add('Only allow past dates', (() => (
+    <Container>
+      <DatePicker
+        inputId="date-picker-id"
+        initialDate={ null }
+        isFocused={ boolean('isFocused', true) }
+        isFutureDateSelectable={ boolean('isFutureDateSelectable', false) }
+        isPastDateSelectable={ boolean('isPastDateSelectable', true) }
+        onDateChange={ action('onDateChange') }
+      />
+    </Container>
+  )))
+  .add('Only allow future dates', (() => (
+    <Container>
+      <DatePicker
+        inputId="date-picker-id"
+        initialDate={ null }
+        isFocused={ boolean('isFocused', true) }
+        isFutureDateSelectable={ boolean('isFutureDateSelectable', true) }
+        isPastDateSelectable={ boolean('isPastDateSelectable', false) }
+        onDateChange={ action('onDateChange') }
+      />
+    </Container>
+  )))
+  .add('With a custom range checker', (() => (
+    <Container>
+      <p>Custom function for
+        <code>  isOutsideRange  </code> prop may look something like
+        <code>  {'{ (day) => {return day.day() % 2 === 0; } }'}</code>
+      </p>
+      <DatePicker
+        inputId="date-picker-id"
+        initialDate={ null }
+        isFocused={ boolean('isFocused', true) }
+        isOutsideRange={ (day) => {return day.day() % 2 === 0; } }
+        onDateChange={ action('onDateChange') }
+      />
+    </Container>
+  )))
+  .add('With a custom initial starting month', (() => (
+    <Container>
+      <p>Adjust the left hand month using a function like:
+        <code>  () => moment().subtract(1, 'months')</code> with the
+        <code>  initialVisibleMonth  </code> prop.
+      </p>
+      <DatePicker
+        inputId="date-picker-id"
+        initialDate={ null }
+        initialVisibleMonth={ () => moment().subtract(1, 'months') }
+        isFocused={ boolean('isFocused', true) }
         onDateChange={ action('onDateChange') }
       />
     </Container>