+++ src/components/DatePicker/DateRangePicker.story.js
@@ -1,3 +1,4 @@
+/* eslint-disable react/jsx-no-bind */
 import React from 'react';
 import styled from 'styled-components';
 import moment from 'moment';
@@ -10,24 +11,24 @@ import DateRangePicker from './DateRangePicker';
 
 const presetPanelOptions = [
   {
-    'label': 'Last 7 days',
-    'startDate': moment().subtract(7, 'days'),
-    'endDate': moment(),
+    label: 'Last 7 days',
+    startDate: moment().subtract(7, 'days').startOf('day'),
+    endDate: moment().endOf('day'),
   },
   {
-    'label': 'Last 30 days',
-    'startDate': moment().subtract(30, 'days'),
-    'endDate': moment(),
+    label: 'Last 30 days',
+    startDate: moment().subtract(30, 'days').startOf('day'),
+    endDate: moment().endOf('day'),
   },
   {
-    'label': 'Last 60 days',
-    'startDate': moment().subtract(60, 'days'),
-    'endDate': moment(),
+    label: 'Last 60 days',
+    startDate: moment().subtract(60, 'days').startOf('day'),
+    endDate: moment().endOf('day'),
   },
   {
-    'label': 'This Month',
-    'startDate': moment().startOf('month'),
-    'endDate': moment().endOf('month'),
+    label: 'This Month',
+    startDate: moment().startOf('month'),
+    endDate: moment().endOf('month'),
   },
 ];
 
@@ -80,13 +81,68 @@ stories
       />
     </Container>
   )))
-  .add('With past dates selectable', (() => (
+  .add('Allow any dates', (() => (
     <Container>
       <DateRangePicker
         endDateInputId='end-date-id-0'
         focusedInput='endDate'
         initialStartDate={ moment() }
+        isFutureDateSelectable={ boolean('isFutureDateSelectable', true) }
         isPastDateSelectable={ boolean('isPastDateSelectable', true) }
+        onDatesChange={ action('onDatesChange') }
+        startDateInputId='start-date-id-0'
+      />
+    </Container>
+  )))
+  .add('Only allow future dates', (() => (
+    <Container>
+      <DateRangePicker
+        endDateInputId='end-date-id-0'
+        focusedInput='endDate'
+        initialStartDate={ moment() }
+        isFutureDateSelectable={ boolean('isFutureDateSelectable', true) }
+        isPastDateSelectable={ boolean('isPastDateSelectable', false) }
+        onDatesChange={ action('onDatesChange') }
+        startDateInputId='start-date-id-0'
+      />
+    </Container>
+  )))
+  .add('Only allow past dates', (() => (
+    <Container>
+      <DateRangePicker
+        endDateInputId='end-date-id-0'
+        focusedInput='endDate'
+        initialVisibleMonth={ () => moment().subtract(1, 'months') }
+        isFutureDateSelectable={ boolean('isFutureDateSelectable', false) }
+        isPastDateSelectable={ boolean('isPastDateSelectable', true) }
+        onDatesChange={ action('onDatesChange') }
+        startDateInputId='start-date-id-0'
+      />
+    </Container>
+  )))
+  .add('With custom range checker', (() => (
+    <Container>
+      <DateRangePicker
+        endDateInputId='end-date-id-0'
+        focusedInput='endDate'
+        initialStartDate={ moment() }
+        isOutsideRange={ action('isOutsideRange') }
+        onDatesChange={ action('onDatesChange') }
+        startDateInputId='start-date-id-0'
+      />
+    </Container>
+  )))
+  .add('With an initial starting month', (() => (
+    <Container>
+      <p>Adjust the left hand month using a function like:
+        <code>  () => moment().subtract(1, 'months')</code> with the
+        <code>  initialVisibleMonth  </code> prop.
+      </p>
+      <DateRangePicker
+        endDateInputId='end-date-id-0'
+        focusedInput='endDate'
+        initialVisibleMonth={ () => moment().subtract(1, 'months') }
+        isOutsideRange={ action('isOutsideRange') }
         onDatesChange={ action('onDatesChange') }
         startDateInputId='start-date-id-0'
       />
@@ -97,7 +153,7 @@ stories
       <DateRangePicker
         endDateInputId='end-date-id-0'
         focusedInput='startDate'
-        keepOpenOnDateSelect={ boolean('keepOpenOndateSelect', true) }
+        keepOpenOnDateSelect={ boolean('keepOpenOnDateSelect', true) }
         onDatesChange={ action('onDatesChange') }
         startDateInputId='start-date-id-0'
       />
@@ -111,6 +167,19 @@ stories
         keepOpenAlways={ boolean('keepOpenAlways', true) }
         onDatesChange={ action('onDatesChange') }
         startDateInputId='start-date-id-0'
+      />
+    </Container>
+  )))
+  .add('With time inputs', (() => (
+    <Container>
+      <DateRangePicker
+        endDateInputId='end-date-id-0'
+        endTimeInputId='endTimeInputId'
+        focusedInput='startDate'
+        hasTimeInputs={ boolean('hasTimeInputs', true) }
+        onDatesChange={ action('onDatesChange') }
+        startDateInputId='start-date-id-0'
+        startTimeInputId='startTimeInputId'
       />
     </Container>
   )))
@@ -160,12 +229,27 @@ stories
       />
     </Container>
   )))
+  .add('With no border and time inputs', (() => (
+    <Container>
+      <DateRangePicker
+        endDateInputId='end-date-id-0'
+        endTimeInputId='endTimeInputId'
+        focusedInput='startDate'
+        isBorderless={ boolean('isBorderless', true) }
+        hasTimeInputs={ boolean('hasTimeInputs', true) }
+        onDatesChange={ action('onDatesChange') }
+        startDateInputId='start-date-id-0'
+        startTimeInputId='startTimeInputId'
+      />
+    </Container>
+  )))
   .add('With preset options and no border on calendar', (() => (
     <Container>
       <DateRangePicker
         endDateInputId='end-date-id-0'
         focusedInput='startDate'
         isBorderless={ boolean('isBorderless', true) }
+        hasTimeInputs={ boolean('hasTimeInputs', false) }
         isPastDateSelectable={ boolean('isPastDateSelectable', true) }
         keepOpenOnDateSelect={ boolean('keepOpenOndateSelect', true) }
         onDatesChange={ action('onDatesChange') }
@@ -196,7 +280,6 @@ stories
       />
     </Container>
   )))
-  /* eslint-disable react/jsx-no-bind */
   .add('With panelButtons as render props function returning an array', (() => (
     <Container>
       <DateRangePicker
@@ -206,7 +289,7 @@ stories
         isPastDateSelectable={ boolean('isPastDateSelectable', true) }
         keepOpenOnDateSelect={ boolean('keepOpenOndateSelect', true) }
         onDatesChange={ action('onDatesChange') }
-        panelButtons={ ({ onDatesChange }) => ([
+        panelButtons={ ({ endDate, onDatesChange, startDate }) => ([
           <Button
             key={ 0 }
             style="plain"
@@ -219,7 +302,7 @@ stories
           <Button
             key={ 1 }
             style="highlight"
-            onClick={ action('onClick Apply') }>
+            onClick={ () => action('onClick Apply')({ startDate, endDate }) }>
             Apply
           </Button>,
         ]) }
@@ -227,7 +310,40 @@ stories
         startDateInputId='start-date-id-0'
       />
     </Container>
-    /* eslint-enable react/jsx-no-bind */
+  )))
+  .add('With hasTimeInput and presetPanelOptions', (() => (
+    <Container>
+      <DateRangePicker
+        endDateInputId='end-date-id-0'
+        endTimeInputId='endTimeInputId'
+        focusedInput='startDate'
+        hasTimeInputs={ boolean('hasTimeInputs', true) }
+        isBorderless={ boolean('isBorderless', true) }
+        isPastDateSelectable={ boolean('isPastDateSelectable', true) }
+        keepOpenOnDateSelect={ boolean('keepOpenOndateSelect', true) }
+        onDatesChange={ action('onDatesChange') }
+        panelButtons={ ({ endDate, onDatesChange, startDate }) => ([
+          <Button
+            key={ 0 }
+            style="plain"
+            onClick={ () => {
+              action('onClick Clear with onDatesChange via render props')();
+              onDatesChange({ startDate: null, endDate: null });
+            } }>
+            Clear
+          </Button>,
+          <Button
+            key={ 1 }
+            style="highlight"
+            onClick={ () => action('onClick Apply')({ startDate, endDate }) }>
+            Apply
+          </Button>,
+        ]) }
+        presetPanelOptions={ presetPanelOptions }
+        startDateInputId='start-date-id-0'
+        startTimeInputId='startTimeInputId'
+      />
+    </Container>
   )));
 
 const Container = styled.div`