+++ src/components/Textarea/Textarea.story.js
@@ -23,6 +23,7 @@ stories
           isDisabled={ boolean('isDisabled', false) }
           defaultValue='Delete this default value and see the placeholder'
           maxLength={ number('maxLength', 250) }
+          numRows={ number('numRows', 25) }
           placeholder={ text('placeHolder', 'Enter a comment') }
           onBlur={ action('Textarea: onBlur') }
           onChange={ action('Textarea: onChange') }